When fetching tags from git do not merge-pull
git pull may change the checkedout tree and update it to the latest
branch available upstream. This is not always the desired action,
when I launch the script I expect that it doesn't change the current
checkedout revision (even if master has been moved upstream) and
mark it with the selected tag.