Do not push all tags, but just the newly created
There may be local tags created previously that the user may not
want to push (and potentially leak private/development branches).

This commit fixes it by pushing only the newly created tag.

Moreover the git remote used for pushing is "origin", but this default
should really be a configurable option.