Properly detecting bash path for all systems
On Linux it's common to have some bash(-compatible) shell available at `/bin/bash`, on pretty much all other systems the default is `/usr/bin/bash` or even `/usr/local/bin/bash` the default (if bash is installed at all). This fixes the detection.

On OS X specifically there is a shell available in `/bin/bash` but it's only bash-like.